Cool things!
Weather, and an interface for it.
We now actually do an API call to forcast.io.
We then parse the returning data, and Toast the temp and cloud cover.

There is a prompt to resume an old inspection, or start a new one.

Turns out that network activity cannot occure on the main thread. So Hive-DAQ is now somewhat multi-threaded! Async tasks happen in the Weather class, this then in turn caused the need for the interface. Anything that uses Weather has to pass itself to it, and then implement the interfaces. It's pretty cool!